Add standalone Python server
The server mirrors the jsdb in a sqlite database. It serves all files (js,
css, images…) without modifications, except for the index, in which it
injects the state of the database at startup.
JSON query can be made against /jsdb, which returns the current jsdb as JSON.
The slite database keep every single revision of an ingredient or recipe, but
only the last revision is considered as valid.
At the moment, it is not possible to update the database.

The init_db script creates the database tables and fills it with samples.